clean up pang's lcm debugging and move to test dir
